ascending order: fixed unwanted mouse input
Changed the instruction of the activity.
A possible bug was found when mouse input was taken into account
even when the animation was still going on. This problem was reflected
when the mouse was Released and the tile went to the previous
location. This was solved by accepting the onPressed action only
when the animation is over

Signed-off by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>

Signed-off-by: Rohit Das <rohit.das950@gmail.com>