ascending_order: sublevel score, ok button, levels
Added sub level score, which will display the current sublevel in which
the player is currently present.

Level changes: Previously on changing the levels from the arrow key,
the current sib level was getting incremented. As of now, on clicking
the left or right arrow keys, the current sub level is reset to zero
and the current level is decremented or incremented accordingly.
On winning, the current sub level is incremented.

The ok button is repositioned above the bar, so that it doesn't get
overriden by the bar when resized

Signed-off by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>

Signed-off-by: Rohit Das <rohit.das950@gmail.com>